roles/system_info/files/requirements.txt
blkinfo==0.2.0
distro==1.8.0
py-cpuinfo==8.0.0
psutil==5.9.2